I rewrote the saving media feature to improve it and fixed the saving video functionality.
To do this, I installed the plugins for saving images and videos from Draft.js so that I could better support the adding images/video functionality. Furthermore, I rewrote all the code in order to implement it in such a way. I created two separate tabs, one for uploading images and one for uploading videos by URL. I created two folders in the repository that were dedicated to these two tabs and added code for 1) Adding the image/video 2) integrating the image/video into a textbox that also allowed the user to write text into the image/video "box". I began to make some edits to backend/index.js to work on saving these to the backend.