Implementation of the backend for saving images.
In this code, I created get and post endpoints in the backend codes for /api/images and a fetchImages and saveImages functions in my frontend code for the image editor in order to have the image editor be associated with users and for the images to save to the backend. I also added a button called "Save Images" which allowed the users to save their finished images to the backend.