Implementation of backend/frontend code for saving videos.
In this commit, I created get and post endpoints in index.js in the backend folder as well as fetchVideos and saveVideos functions in index.js in the VideoApp folder in order to allow the video editor to save text/videos. Each video editor is associated with a userEmail, so that the video editor is preserved whenever hte user logs out and logs back in. I also created a save videos content button on the page to allow the user to save their work.